requirements.txt
streamlit
streamlit-extras
google-generativeai
Pillow
python-dotenv